Move `CustomAsyncImage` component to its own file
